Mutex and signal handling added
Only 1 thread can be sending on the socket at a time as it is locked with the mutex lock. Added basic signal handling so the C++ program runs until it is stopped. Messages are very grouped, but final implementation will probably be event driven anyway, might look into semaphore stuff if worthwhile.