feat: Killswitch, improved client, added debug messages
Added a kill switch that can be activated with the respective button on
the client. This is like a pause button for the network communications
of the server. When red, no network comms (doesnt send or receive
messages), when green, can properly talk to other nodes.

Added an 'Action' section that can be used to see the action that each
node is taking. At the moment it records state changes, messages sent
and messages received, can easily be exanded throughout development.

Adjusted the client so that more info could be displayed, data input is
now on the bottom right, exactly the same functionality.

Also changed the time to be based on the first received server
communication rather than raw Unix epoch time.