Feat: Leader Election/AppendEntries
Added leader election, so now when the c++ program is run a leader is
elected (election timeout, becomes candidate, gathers majority of votes, becomes leader), and all other servers are followers.

Needed to implement AppendEntries to get leader election to work, this
follows a very similar format to the request vote messages.