feat: implemented log replication
Required some changes with temporary implementation of AppendEntries RPC used for heatbeats.

Some things added that aren't outlined in spec and some things not used, for example `matchIndex` as it's not explained how it's supposed to be used (or why it's useful).

Required additional server functionality to send messages to servers by ID from outside of the server class - each server gets given a unique AppendEntries message.